Perform black button click at (re)initialization.
This shows the black color as active. However, the previously chosen color is not kept with an update. But, I couldn't figure out yet how to load the color after the spinner initialization. Maybe it would be easier to make the entries in the matrix Enum and possibly mark them as hints and editable.